feat: add backend example to Home. add proper manifest
