+++ packages/pole/public/manifest.json
@@ -1,9 +1,9 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "E-Vote",
+  "name": "Local Election facilitator",
   "icons": [
     {
-      "src": "favicon.ico",
+      "src": "ballot-box.png",
       "sizes": "64x64 32x32 24x24 16x16",
       "type": "image/x-icon"
     }
